feat(api): BotSocialRelationsContainer
分离bot的社交关系相关api为独立的容器接口
